Fix 'required' for non-top-level fields
Fixes: project-radius/radius#3688

This change fixes the required flag for non-top-level fields. We had
code for this and it was not doing it's job. We also had a test for
this, but were verifying the wrong behavior.

The bug was that we were mishandling two instances of the schema, the type
schema which contains the list of required fields, and the property
schema that contains the definition of the field. We were looking at the
property schema for the *is required* check when we should have looked
at the type schema.

This change also includes schema updates to ~200 files for the 100 types
that were bugged.